In 1219, he went to Egypt in an attempt to convert the sultan al-Kamil and put an end to the conflict of the Fifth Crusade.[4] In 1223, he arranged for the first live nativity scene as part of the annual Christmas celebration in Greccio.[b][5][6] According to Christian tradition, in 1224 he received the stigmata during the apparition of a Seraphic angel in a religious ecstasy.[7]
He founded the men's Order of Friars Minor, the women's Order of St. Clare, the Third Order of St. Francis and the Custody of the Holy Land. Once his community was authorized by the Pope, he withdrew increasingly from external affairs.

Francis is associated with patronage of animals and the environment. It became customary for churches to hold ceremonies blessing animals on his feast day of 4 October. He is known for devotion to the Eucharist.[8] Along with Catherine of Siena, he was designated patron saint of Italy.